# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Server Inventory System API
  description: |
    REST API for the Server Inventory System.

    Manages server asset inventory and provides access to VMware vSphere and
    New Relic sync data. All endpoints require JWT authentication.

    **Authentication:** Pass a JWT token in the `Authorization: Bearer <token>` header.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://inventory.example.com
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/vsphere:
    get:
      tags: [vSphere]
      summary: List vSphere records
      operationId: listVsphere
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of vSphere records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VsphereRecordPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/vsphere/sync:
    post:
      tags: [vSphere]
      summary: Trigger manual vSphere sync
      description: Immediately starts a vSphere sync job, ignoring the configured schedule.
      operationId: triggerVsphereSync
      responses:
        '202':
          description: Sync job accepted and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: A sync job is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/vsphere/sync/status:
    get:
      tags: [vSphere]
      summary: Get vSphere sync status
      description: Returns the status and result of the most recent vSphere sync run.
      operationId: getVsphereSyncStatus
      responses:
        '200':
          description: vSphere sync status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/vsphere/{hostname}:
    parameters:
      - $ref: '#/components/parameters/HostnameParam'

    get:
      tags: [vSphere]
      summary: Get vSphere record by hostname
      operationId: getVsphereByHostname
      responses:
        '200':
          description: vSphere record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VsphereRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ------------------------------------------------------------
  # New Relic
  # ------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ------------------------------------------------------------
  # Parameters
  # ------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page number
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
    HostnameParam:
      name: hostname
      in: path
      required: true
      description: Asset hostname (primary key)
      schema:
        type: string
      example: web-prod-01
...
  schemas:
...
    VsphereRecord:
      type: object
      properties:
        hostname:
          type: string
          example: web-prod-01
        fqdn:
          type: string
          nullable: true
          example: web-prod-01.corp.example.com
        vm_name:
          type: string
          example: web-prod-01
        vm_id:
          type: string
          description: vSphere managed object ID (moId)
          example: vm-1042
        cluster:
          type: string
          nullable: true
          example: prod-cluster-01
        datacenter:
          type: string
          nullable: true
          example: DC1
        datastore:
          type: string
          nullable: true
          example: ds-ssd-01
        cpu_count:
          type: integer
          nullable: true
          example: 4
        cpu_cores:
          type: integer
          nullable: true
          example: 2
        memory_mb:
          type: integer
          nullable: true
          example: 8192
        memory_gb:
          type: integer
          nullable: true
          example: 8
        disk_gb:
          type: integer
          nullable: true
          example: 100
        power_state:
          type: string
          enum: [poweredOn, poweredOff, suspended]
          nullable: true
          example: poweredOn
        guest_os:
          type: string
          nullable: true
          example: "Ubuntu Linux (64-bit)"
        tools_status:
          type: string
          enum: [toolsOk, toolsOld, toolsNotRunning, toolsNotInstalled]
          nullable: true
          example: toolsOk
        ipv4_address:
          type: string
          nullable: true
          example: 10.0.1.50
        ipv6_address:
          type: string
          nullable: true
          example: null
        last_synced_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-01T02:05:00Z"
        created_at:
          type: string
          format: date-time
          example: "2026-02-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-03-01T02:05:00Z"

    # -- New Relic --
...
    VsphereRecordPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/VsphereRecord'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        total_elements:
          type: integer
          example: 95
        total_pages:
          type: integer
          example: 5
...
    SyncStatus:
      type: object
      properties:
        service:
          type: string
          enum: [vsphere, newrelic]
          example: vsphere
        status:
          type: string
          enum: [idle, running, success, failed]
          example: success
        last_run_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-02T02:05:00Z"
        records_synced:
          type: integer
          nullable: true
          example: 87
        message:
          type: string
          nullable: true
          description: Error detail when status is `failed`
          example: null
    SyncAccepted:
      type: object
      properties:
        service:
          type: string
          enum: [vsphere, newrelic]
          example: vsphere
        message:
          type: string
          example: Sync job started

    # -- Error --
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: "Asset 'web-prod-01' not found"
        timestamp:
          type: string
          format: date-time
          example: "2026-03-02T10:00:00Z"
        path:
          type: string
          example: /api/inventory/web-prod-01

  # ------------------------------------------------------------
  # Reusable responses
  # ------------------------------------------------------------
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'